[Oracle] Refactor code
