+++ flink-cdc-connect/flink-cdc-source-connectors/flink-connector-oracle-cdc/src/main/java/com/ververica/cdc/connectors/oracle/source/reader/fetch/OracleSourceFetchTaskContext.java
@@ -27,6 +27,7 @@ import com.ververica.cdc.connectors.base.source.meta.split.SourceSplitBase;
 import com.ververica.cdc.connectors.base.source.reader.external.JdbcSourceFetchTaskContext;
 import com.ververica.cdc.connectors.base.utils.SourceRecordUtils;
 import com.ververica.cdc.connectors.oracle.source.config.OracleSourceConfig;
+import com.ververica.cdc.connectors.oracle.source.handler.OracleSchemaChangeEventHandler;
 import com.ververica.cdc.connectors.oracle.source.meta.offset.RedoLogOffset;
 import com.ververica.cdc.connectors.oracle.source.utils.OracleUtils;
 import com.ververica.cdc.connectors.oracle.util.ChunkUtils;
@@ -65,7 +66,6 @@ import org.slf4j.LoggerFactory;
 
 import java.sql.SQLException;
 import java.time.Instant;
-import java.util.HashMap;
 import java.util.Map;
 
 import static com.ververica.cdc.connectors.oracle.source.utils.OracleConnectionUtils.createOracleConnection;
@@ -145,13 +145,7 @@ public class OracleSourceFetchTaskContext extends JdbcSourceFetchTaskContext {
                         DataChangeEvent::new,
                         metadataProvider,
                         schemaNameAdjuster,
-                        event -> {
-                            Map<String, Object> source = new HashMap<>();
-                            Struct sourceInfo = event.getSource();
-                            String scn = sourceInfo.getString(SCN_KEY);
-                            source.put(SCN_KEY, scn);
-                            return source;
-                        });
+                        new OracleSchemaChangeEventHandler());
 
         final OracleChangeEventSourceMetricsFactory changeEventSourceMetricsFactory =
                 new OracleChangeEventSourceMetricsFactory(
