+++ flink-connector-oracle-cdc/src/main/java/com/ververica/cdc/connectors/oracle/source/handler/OracleSchemaChangeEventHandler.java
@@ -1,0 +1,38 @@
+/*
+ * Copyright 2023 Ververica Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.ververica.cdc.connectors.oracle.source.handler;
+
+import com.ververica.cdc.connectors.base.relational.handler.SchemaChangeEventHandler;
+import io.debezium.schema.SchemaChangeEvent;
+import org.apache.kafka.connect.data.Struct;
+
+import java.util.HashMap;
+import java.util.Map;
+
+import static io.debezium.connector.oracle.SourceInfo.SCN_KEY;
+
+public class OracleSchemaChangeEventHandler implements SchemaChangeEventHandler {
+
+    @Override
+    public Map<String, Object> parseSource(SchemaChangeEvent event) {
+        Map<String, Object> source = new HashMap<>();
+        Struct sourceInfo = event.getSource();
+        String scn = sourceInfo.getString(SCN_KEY);
+        source.put(SCN_KEY, scn);
+        return source;
+    }
+}
